+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.3
 - Fixed stack-crafting in the Wireless Crafting Grid (crafting table) causing lag on a dedicated server (raoulvdberge)
+- Refined Storage 1.5.33 compatibility (raoulvdberge)
 
 ### 0.2
 - Updated Forge to 2493 (MC 1.12.2) (raoulvdberge)